Get R/G/B data from images
Code by Nathan Hearnsberger on separating images of Airbnb listings into RGB components. Idea is that the brighter the picture, the more attractive the picture might be, and the more appealing the room seems. Given more time in the competition, we would have also tried training neural networks to recognize specific items in the rooms